rename `changed_files` config setting to `changed_proc`
This is prep for having a non-singleton config handler.  This renames
the `changed_files` setting to make it be named more consistent with
the other proc settings.

This also moves the config class to its own file, again, in prep for
a non-singleton config.
